refactor: move shipped-proposals to ts
